requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
arrow==0.12.0
asn1crypto==0.23.0        # via cryptography
attrs==17.3.0             # via pytest
babel==2.5.1              # via delorean
billiard==3.5.0.4         # via celery
celery==4.2.1
certifi==2017.11.5        # via requests
cffi==1.11.2              # via cryptography
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coverage==4.4.2
cryptography==2.1.4       # via requests-ntlm
daemonize==2.4.7
defusedxml==0.5.0         # via python3-openid
delorean==0.6.0
django-allauth==0.34.0
django-anymail==1.2
django-appconf==1.0.1     # via django-image-cropping
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
django-grappelli==2.10.1
django-guardian==1.4.9
django-helusers==0.4.2
django-hstore==1.4.2
django-image-cropping==1.1.0
django-jinja==2.4.1
django-modeltranslation==0.12.1
django-mptt==0.8.7        # via django-munigeo
django-munigeo==0.2.0
django-parler==1.8.1
django-reversion==2.0.11
django==1.11.11
djangorestframework-jwt==1.11.0
djangorestframework==3.7.3
drf-oidc-auth==0.9        # via django-helusers
easy-thumbnails==2.5
fake-factory==9999.9.9
first==2.0.1              # via pip-tools
flake8==3.5.0
freezegun==0.3.9
future==0.16.0            # via pyjwkest
humanize==0.5.1           # via delorean
icalendar==4.0.0
idna==2.6                 # via cryptography, requests
iso8601==0.1.12
jinja2==2.10
kombu==4.2.1              # via celery
lxml==4.1.1
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
ntlm-auth==1.0.6          # via requests-ntlm
oauthlib==2.0.6           # via requests-oauthlib
olefile==0.44             # via pillow
pillow==4.3.0             # via easy-thumbnails
pip-tools==1.11.0
pluggy==0.6.0             # via pytest
psycopg2==2.7.3.2
py==1.5.2                 # via pytest
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pycryptodomex==3.4.7      # via pyjwkest
pyflakes==1.6.0           # via flake8
pyjwkest==1.4.0           # via drf-oidc-auth
pyjwt==1.5.3
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.3.0
python-dateutil==2.6.1    # via arrow, delorean, freezegun, icalendar
python-docx==0.8.6
python3-openid==3.1.0     # via django-allauth
pytz==2018.3
pyyaml==3.12              # via django-munigeo
raven==6.3.0
requests-cache==0.4.13    # via django-munigeo
requests-ntlm==1.1.0
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.4
six==1.11.0
tzlocal==1.5.1            # via delorean
urllib3==1.22             # via requests
vine==1.1.4               # via amqp
xlsxwriter==1.0.2
